+++ src/common-ui/src/main/java/ee/ria/xroad/common/request/ManagementRequestSender.java
@@ -22,19 +22,13 @@
  */
 package ee.ria.xroad.common.request;
 
-import static ee.ria.xroad.common.ErrorCodes.X_HTTP_ERROR;
-import static ee.ria.xroad.common.ErrorCodes.X_INTERNAL_ERROR;
-import static ee.ria.xroad.common.util.AbstractHttpSender.CHUNKED_LENGTH;
-import static ee.ria.xroad.common.util.MimeUtils.TEXT_XML_UTF8;
-import static ee.ria.xroad.common.util.MimeUtils.getBaseContentType;
-import static org.eclipse.jetty.http.MimeTypes.TEXT_XML;
-
 import java.io.ByteArrayInputStream;
 import java.io.InputStream;
 import java.net.URI;
 
 import javax.xml.soap.SOAPException;
 
+import lombok.extern.slf4j.Slf4j;
 import org.w3c.dom.Node;
 import org.w3c.dom.NodeList;
 
@@ -50,7 +44,13 @@ import ee.ria.xroad.common.message.SoapMessageImpl;
 import ee.ria.xroad.common.message.SoapParserImpl;
 import ee.ria.xroad.common.message.SoapUtils;
 import ee.ria.xroad.common.util.HttpSender;
-import lombok.extern.slf4j.Slf4j;
+
+import static ee.ria.xroad.common.ErrorCodes.X_HTTP_ERROR;
+import static ee.ria.xroad.common.ErrorCodes.X_INTERNAL_ERROR;
+import static ee.ria.xroad.common.util.AbstractHttpSender.CHUNKED_LENGTH;
+import static ee.ria.xroad.common.util.MimeTypes.TEXT_XML;
+import static ee.ria.xroad.common.util.MimeTypes.TEXT_XML_UTF8;
+import static ee.ria.xroad.common.util.MimeUtils.getBaseContentType;
 
 /**
  * Sends various management requests. Authentication certificate registration