+++ src/common-util/src/main/java/ee/ria/xroad/common/message/SoapMessageDecoder.java
@@ -22,7 +22,12 @@
  */
 package ee.ria.xroad.common.message;
 
-import ee.ria.xroad.common.CodedException;
+import java.io.Closeable;
+import java.io.IOException;
+import java.io.InputStream;
+import java.util.HashMap;
+import java.util.Map;
+
 import lombok.extern.slf4j.Slf4j;
 import org.apache.james.mime4j.MimeException;
 import org.apache.james.mime4j.parser.AbstractContentHandler;
@@ -31,18 +36,12 @@ import org.apache.james.mime4j.stream.BodyDescriptor;
 import org.apache.james.mime4j.stream.Field;
 import org.apache.james.mime4j.stream.MimeConfig;
 
-import java.io.Closeable;
-import java.io.IOException;
-import java.io.InputStream;
-import java.util.HashMap;
-import java.util.Map;
+import ee.ria.xroad.common.CodedException;
 
 import static ee.ria.xroad.common.ErrorCodes.*;
-import static ee.ria.xroad.common.util.MimeTypes.MULTIPART_RELATED;
-import static ee.ria.xroad.common.util.MimeTypes.XOP_XML;
+import static ee.ria.xroad.common.util.MimeTypes.*;
 import static ee.ria.xroad.common.util.MimeUtils.HEADER_CONTENT_TYPE;
 import static ee.ria.xroad.common.util.MimeUtils.getBaseContentType;
-import static org.eclipse.jetty.http.MimeTypes.TEXT_XML;
 
 /**
  * Decodes SOAP messages from an input stream.