+++ src/common-util/src/main/java/ee/ria/xroad/common/util/MimeTypes.java
@@ -23,10 +23,9 @@
 package ee.ria.xroad.common.util;
 
 /**
- * Extends jetty MimeTypes class, adding some other mime type constants
- * used in MessageEncoder/Decoder.
+ * Mime types class containing mime type constants used in MessageEncoder/Decoder and other places.
  */
-public class MimeTypes extends org.eclipse.jetty.http.MimeTypes {
+public final class MimeTypes {
 
     public static final String OCSP_REQUEST = "application/ocsp-request";
     public static final String OCSP_RESPONSE = "application/ocsp-response";
@@ -43,8 +42,7 @@ public class MimeTypes extends org.eclipse.jetty.http.MimeTypes {
 
     public static final String GZIP = "application/gzip";
 
-    public static final String HASH_CHAIN_RESULT =
-            "application/hash-chain-result";
+    public static final String HASH_CHAIN_RESULT = "application/hash-chain-result";
 
     public static final String HASH_CHAIN = "application/hash-chain";
 
@@ -52,4 +50,16 @@ public class MimeTypes extends org.eclipse.jetty.http.MimeTypes {
     public static final String JSON_RPC = "application/json-rpc";
 
     public static final String XOP_XML = "application/xop+xml";
+
+    public static final String TEXT_PLAIN = "text/plain";
+    public static final String TEXT_PLAIN_UTF8 = MimeUtils.contentTypeWithCharset(TEXT_PLAIN, MimeUtils.UTF8);
+
+    public static final String TEXT_XML = "text/xml";
+    public static final String TEXT_XML_UTF8 = MimeUtils.contentTypeWithCharset(TEXT_XML, MimeUtils.UTF8);
+
+    public static final String TEXT_HTML = "text/html";
+    public static final String TEXT_HTML_UTF8 = MimeUtils.contentTypeWithCharset(TEXT_HTML, MimeUtils.UTF8);
+
+    private MimeTypes() {
+    }
 }